+++ src/App.tsx
@@ -17,16 +17,13 @@ const App: React.FC = () => {
         <Route key={key} path={question.link} element={<BaseQuestionPage question={question} />} />;
 
     return <Router>
-        <div className="transition-all duration-200 bg-gray-200 min-h-screen w-full">
-            <Routes>
-                {questions.map(RenderQuestion)}
-                <Route path="/search/expanded" element={<ExpandedSearchItem />} />
-                <Route path="/search" element={<Search />} />
-                <Route path="/map" element={<Map />} />
-                <Route path="*" element={<LandingPage />} />
-            </Routes>
-        </div>
-        
+        <Routes>question
+            {questions.map(RenderQuestion)}
+            <Route path="/search/expanded" element={<ExpandedSearchItem />} />
+            <Route path="/search" element={<Search />} />
+            <Route path="/map" element={<Map />} />
+            <Route path="*" element={<LandingPage />} />
+        </Routes>
     </Router>;
 };
